// TS/src/types/other.ts
export type CounterType = {
  title: string
  count: number
  icon?: IconType
  variant?: string
  prefix?: string
  suffix?: string
}